Implement permission fixes throughout platform privapp permissions
"These permissions fall within the platform directory, When adding this line:
ro.control_privapp_permissions=log
to the system/product/build.prop, the device booted, but fired errors saying
which apps didn't have proper permissions. Without this line, people would
have problems booting because the system would go in an infinite loop of
not booting

These are the fixed permissions that provides a hotfix for the user. Maintainer
should add these permissions to a different file in the device_google_marlin
tree.

Permissions changed in:
 com.liquid.ota
 com.android.settings
 com.android.systemui

Signed-off-by: kingbri <bdashore3@gmail.com>